+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## [Unreleased]
 
+## [0.10.0] - 2025-07-28
+
 ### Added
 
 - Add `feature-into-crate` command that only inserts feature documentation into crate documentation 
@@ -136,7 +138,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2025-07-09
 
 <!-- next-url -->
-[Unreleased]: https://github.com/bluurryy/cargo-insert-docs/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/bluurryy/cargo-insert-docs/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/bluurryy/cargo-insert-docs/releases/tag/v0.10.0
 [0.9.0]: https://github.com/bluurryy/cargo-insert-docs/releases/tag/v0.9.0
 [0.8.0]: https://github.com/bluurryy/cargo-insert-docs/releases/tag/v0.8.0
 [0.7.0]: https://github.com/bluurryy/cargo-insert-docs/releases/tag/v0.7.0